+++ src/ir/mod.rs
@@ -3,5 +3,5 @@ pub mod schema;
 pub mod types;
 
 pub use query::{NativeListBind, Parameter, Query, QueryCmd, ResultColumn};
-pub use schema::{Column, ScalarFunction, Schema, Table};
+pub use schema::{Column, ScalarFunction, Schema, Table, TableKind};
 pub use types::SqlType;